Migrate dragged playground to TS
