--- packages/draggable/playgrounds/dflex-react-draggable/src/Lists/List.css
@@ -1,40 +1,0 @@
-.Huge {
-  background: rebeccapurple;
-  height: 800px;
-}
-.Container {
-  display: flex;
-  width: 41.53rem;
-  /* border: 4px green solid; */
-  background: #07718f;
-}
-
-ul {
-  /* display: flex; */
-  flex-direction: column;
-  list-style: none;
-  margin: 10px;
-  padding: 8px;
-  border: 4px #bac5c7 solid;
-  background: #cecece;
-  height: fit-content;
-}
-
-li {
-  position: relative;
-  width: 150px;
-  height: 30px;
-  box-shadow: none;
-  margin-bottom: 8px;
-  margin-top: 4px;
-  user-select: none;
-  color: rgb(9, 30, 66);
-  background: whitesmoke;
-  border-radius: 2px;
-  border-width: 2px;
-  border-style: solid;
-  border-image: initial;
-  border-color: transparent;
-  padding: 8px;
-  transition: transform 0.1s cubic-bezier(0.2, 0, 0, 1) 0s;
-}
