--- packages/draggable/playgrounds/dflex-react-draggable/src/Lists/components/Container.jsx
@@ -1,64 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import React from "react";
-
-import { Draggable } from "@dflex/draggable";
-
-const Container = ({
-  component: ContainerComponent = "div",
-  children,
-  ...rest
-}) => {
-  let mouseEvents;
-
-  let draggable;
-
-  let draggedID = null;
-
-  const onMouseUp = () => {
-    if (draggedID) {
-      mouseEvents.forEach(({ evType, evTarget, handler }) => {
-        evTarget.removeEventListener(evType, handler);
-      });
-
-      draggable.endDragging();
-    }
-  };
-
-  const onMouseMove = (e) => {
-    const { clientX, clientY } = e;
-
-    draggable.dragAt(clientX, clientY);
-  };
-
-  const onMouseDown = (e) => {
-    const { target, button, clientX, clientY } = e;
-
-    // avoid right mouse click and ensure id
-    if (typeof button === "number" && button === 0) {
-      const { id } = target;
-
-      if (id) {
-        draggedID = id;
-
-        draggable = new Draggable(id, { x: clientX, y: clientY });
-
-        mouseEvents = [
-          { evType: "mousemove", evTarget: document, handler: onMouseMove },
-          { evType: "mouseup", evTarget: document, handler: onMouseUp },
-        ];
-
-        mouseEvents.forEach(({ evType, evTarget, handler }) => {
-          evTarget.addEventListener(evType, handler);
-        });
-      }
-    }
-  };
-
-  return (
-    <ContainerComponent onMouseDown={onMouseDown} {...rest}>
-      {children}
-    </ContainerComponent>
-  );
-};
-
-export default Container;
