--- packages/draggable/playgrounds/dflex-react-draggable/src/Lists/components/index.js
@@ -1,4 +1,0 @@
-import Container from "./Container";
-import Draggable from "./Draggable";
-
-export { Container, Draggable };
