--- packages/draggable/playgrounds/dflex-react-draggable/src/Lists/index.js
@@ -1,3 +1,0 @@
-import Lists from "./Lists";
-
-export default Lists;
